# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Localfluence API - influencer marketplace
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listMyRequests
      tags: [requests]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected, completed]
      responses:
        "200":
          description: Collaboration requests for current user
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CollaborationRequest" }
    post:
      operationId: createRequest
      tags: [requests]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRequestBody" }
      responses:
        "200":
          description: Created request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CollaborationRequest" }
  /requests/{id}/status:
    put:
      operationId: updateRequestStatus
      tags: [requests]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateRequestStatusBody" }
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CollaborationRequest" }
...
components:
  schemas:
...
    RequestStatus:
      type: string
      enum: [pending, accepted, rejected, completed]
    CollaborationRequest:
      type: object
      required: [id, businessId, influencerId, status, createdAt]
      properties:
        id: { type: integer }
        businessId: { type: integer }
        businessName: { type: string }
        influencerId: { type: integer }
        influencerName: { type: string }
        campaignId: { type: integer }
        campaignTitle: { type: string }
        message: { type: string }
        status: { $ref: "#/components/schemas/RequestStatus" }
        createdAt: { type: string, format: date-time }
    CreateRequestBody:
      type: object
      required: [influencerId]
      properties:
        influencerId: { type: integer }
        campaignId: { type: integer }
        message: { type: string }
    UpdateRequestStatusBody:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/RequestStatus" }
...