# lib/api-zod/src/generated/types/createRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Localfluence API - influencer marketplace
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRequestBody {
  influencerId: number;
  campaignId?: number;
  message?: string;
}